% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DID.R
\name{DID}
\alias{DID}
\title{Difference-in-Difference Estimator when multiple units are missing.}
\usage{
DID(M, mask)
}
\arguments{
\item{M}{Matrix of observed entries. The input should be N (number of units) by T (number of time periods).}

\item{mask}{Binary mask with the same shape as M containing observed entries.}
}
\value{
The matrix with all missing entries filled.
}
\description{
Difference-in-Difference Estimator when multiple units are missing.
}
\examples{
DID(matrix(c(1,2,3,4),2,2), matrix(c(1,1,1,0),2,2))
}
