% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_estimators.R
\name{imputation_ATT}
\alias{imputation_ATT}
\title{Imputation estimators for ATT}
\usage{
imputation_ATT(
  df,
  unit_id,
  time_id,
  treatment,
  outcome,
  imputation_methods = c("did", "dfm", "sdfm", "knn", "kknn", "mc", "sc", "env", "enh"),
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{unit_id}{unit identifier}

\item{time_id}{time identifier}

\item{treatment}{treatment identifier}

\item{outcome}{outcome identifier}

\item{imputation_methods}{list of methods to pass to imputationY}

\item{...}{additional arguments to pass to imputationY function}
}
\value{
event study coefficients from each method
}
\description{
meta-function to perform reshape and fit multiple imputation methods
}
