% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_estimators.R
\name{imputationY}
\alias{imputationY}
\title{Imputation Estimators for Panel Data}
\usage{
imputationY(
  Y0,
  methods = c("dfm", "sdfm", "kknn", "knn", "did", "mc", "sc", "env", "enh"),
  dfm_args = list(r = 3, p = 1),
  sdfm_args = list(r = 3, q = 1),
  knn_args = list(k = 5),
  kknn_args = list(kernel = "rbfdot", kpar = list(sigma = 0.1), features = 5, k = 3)
)
}
\arguments{
\item{Y0}{N X T matrix with missing values (must be either simultaneous or staggered adoption)}

\item{methods}{character vectors with names of estimators. Must be a subset of ("dfm", "knn", "kknn", "nuclear", "optspace", "softimpute", "hardimpute", "mc", "sc", "env", "enh"). Defaults to all of the above}

\item{dfm_args}{arguments to pass to dfms call (must match dfms::DFM function documentation)}

\item{sdfm_args}{arguments to pass to sparseDFM call}

\item{knn_args}{arguments to pass to filling::KNNimpute call}

\item{kknn_args}{arguments to pass to kernlab::kpca call}
}
\value{
list with imputations corresponding with methods
}
\description{
Imputes missing elements in a N X T panel data matrix and returns a list of all imputations
}
