% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_imputation.R
\name{KKNN}
\alias{KKNN}
\title{Kernel KNN imputation}
\usage{
KKNN(
  Y0,
  kpca_args = list(kernel = "rbfdot", kpar = list(sigma = 0.1), features = 5, k = 3)
)
}
\arguments{
\item{Y0}{N X T matrix with missing entries for treated periods (output from panelMat).}

\item{k}{scalar number of neighbours to average over}

\item{kern}{kernel for kernel PCA}

\item{nf}{number of factors in kernel PCA to match over}
}
\value{
complete matrix with weighted Kernel-KNN imputation entries
}
\description{
Kernel KNN imputation
}
