% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcnnm_fit}
\alias{mcnnm_fit}
\title{This function trains a model for a specified value of lambda_L. The result contains L, u, v, and lambda_L. 
The only difference with mcnnm is that this function trains only for a specified value of lambda_L.}
\usage{
mcnnm_fit(M, mask, lambda_L, to_estimate_u = 1L, to_estimate_v = 1L,
  niter = 1000L, rel_tol = 1e-05, is_quiet = 1L)
}
\arguments{
\item{M}{Matrix of observed entries. The input should be N (number of units) by T (number of time periods).}

\item{mask}{Binary mask with the same shape as M containing observed entries.}

\item{lambda_L}{Required parameter for fitting the model as this function computes the result for a specified value of lambda_L.}

\item{to_estimate_u}{Optional boolean input for wheter estimating fixed unit effects (row means of M) or not. Default is 1.}

\item{niter}{Optional parameter on the number of iterations taken in the algorithm for each fixed value of lambda_L. The default value is 1000 and it is sufficiently large as the algorithm is using warm-start strategy.}

\item{rel_tol}{Optional parameter on the stopping rule. Once the relative improve in objective value drops below rel_tol, execution is halted. Default value is 1e-5.}

\item{is_quiet}{Optional boolean input which indicates whether to print the status of learning and convergence results for Cyclic Coordinate Descent algorithm or not. The default value is 1 (no output is printed).}
}
\value{
The fitted model for the given value of lambda_L.
}
\description{
This function trains a model for a specified value of lambda_L. The result contains L, u, v, and lambda_L. 
The only difference with mcnnm is that this function trains only for a specified value of lambda_L.
}
\examples{
mcnnm_fit(M = replicate(5,rnorm(5)), mask = matrix(rbinom(5*5,1,0.8),5,5), lambda_L = 0.5)
}
\seealso{
\code{\link{mcnnm}}
}
