% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_estimators.R
\name{plot.imputationATT}
\alias{plot.imputationATT}
\title{plot method for imputationATT}
\usage{
plot.imputationATT(x, est_names = colnames(x$event_study_estimates), prec = 3)
}
\description{
plot method for imputationATT
}
