% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_mc}
\alias{predict_mc}
\title{predict complete data matrix from matrix factorization model}
\usage{
predict_mc(m, unit_fe = TRUE, time_fe = TRUE)
}
\arguments{
\item{m}{mcnnm model fit}

\item{unit_fe}{boolean for whether to include unit F (default TRUE)}

\item{time_fe}{boolean for whether to include time FE (default TRUE)}
}
\value{
predicted matrix Y(0)
}
\description{
predict complete data matrix from matrix factorization model
}
