% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi.R
\name{simul_adapt}
\alias{simul_adapt}
\title{This function models simultaneuous adaption and produces the desired binary mask.}
\usage{
simul_adapt(M, N_t, T0, treat_indices = 0)
}
\arguments{
\item{M}{Matrix of observed entries. The input should be N (number of units) by T (number of time periods).}

\item{N_t}{Number of treated units desired.}

\item{T0}{The time just before treatment for all treated units. For instance, if T0 = 2, then first two entries of treated units are counted as control and the rest are treated.}

\item{treat_indices}{Optional indices for treated units. The default is sampling N_t unit from all N units randomly. However, user can manually set some units as treated.}
}
\value{
The masked matrix which is one for control units and treated units before treatment and zero for treated units after treatment.
}
\description{
This function models simultaneuous adaption and produces the desired binary mask.
}
\examples{
simul_adapt(M = replicate(5,rnorm(5)), N_t = 3, T0 = 3)
}
