% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi.R
\name{stag_adapt}
\alias{stag_adapt}
\title{This function models staggered adaption and produces the desired binary mask.}
\usage{
stag_adapt(M, N_t, T0, treat_indices = 0)
}
\arguments{
\item{M}{Matrix of observed entries. The input should be N (number of units) by T (number of time periods).}

\item{N_t}{Number of treated units desired.}

\item{T0}{The first treatment time. The rest of treatment times are equally spaced between T0 to T.}

\item{treat_indices}{Optional indices for treated units. The default is sampling N_t unit from all N units randomly. However, user can manually set some units as treated. Note that indices should be sorted increasingly based on their T0.}
}
\description{
This function models staggered adaption and produces the desired binary mask.
}
\examples{
stag_adapt(M = replicate(5,rnorm(5)), N_t = 3, T0 = 3)
}
